+++ graphql/resolvers/index.js
@@ -1,0 +1,6 @@
+// import resolvers
+const userResolvers = require('./user/userSignup');
+
+module.exports = {
+    ...userResolvers
+}
