+++ graphql/schemas/index.js
@@ -1,0 +1,6 @@
+const { buildSchema } = require('graphql');
+
+// import graphql schemas
+const userSignup = require('./user/userSignup');
+
+module.exports = buildSchema(`${userSignup}`)
