--- schemas/scheme.js
@@ -1,1 +1,0 @@
-const { buildSchema } = require('graphql');
